i should probably unify these build files
