+++ .gitignore
@@ -1,6 +1,6 @@
 dist
 node_modules
-!scripts/nix/dist
+result
 
 *.exe
 vencord_installer
@@ -8,6 +8,7 @@ vencord_installer
 .idea
 
 yarn.lock
+!scripts/nix/yarn.lock
 package-lock.json
 
 *.log
