+++ scripts/nix/patches/index.js
@@ -1,2 +1,3 @@
 require("../dist/patcher.js");
 require("../app.asar");
+
